style: :art: hook.ts logic.tsをファイル分割
