--- src/components/atoms/Button.module.scss
@@ -1,12 +1,0 @@
-.button {
-  justify-content: center;
-  width: 100%;
-  height: 36px;
-  padding: 5px;
-  font-size: 16px;
-  color: var(--color-cream);
-  text-align: center;
-  background-color: var(--color-black);
-  border: solid 1px var(--color-gray);
-  border-radius: 5px;
-}
