+++ src/components/atoms/CustomLink.module.scss
@@ -3,7 +3,7 @@
   height: 16px;
   margin: 24px auto;
   font-size: 14px;
-  color: var(--color-gray);
+  color: var(--color-emerald);
   text-align: center;
   text-decoration: underline;
 }
