+++ src/features/auth/sign-up/logics/validatePassword.ts
@@ -1,0 +1,1 @@
+export const validatePassword = (password: string, passwordConfirm: string) => password === passwordConfirm;
