+++ src/infrastructure/mui/theme.ts
@@ -49,6 +49,13 @@ const theme = createTheme({
         },
       },
     },
+    MuiToggleButton: {
+      styleOverrides: {
+        root: {
+          fontSize: 17,
+        },
+      },
+    },
   },
 });
